+++ src/workflows-app/components/ImportGithub.js
@@ -4,16 +4,10 @@ import { ButtonPrimary } from 'src/components/common';
 import { icon } from 'src/components/icons';
 import { ValidatedInput } from 'src/components/input';
 import { TooltipCell } from 'src/components/table';
-import { Ajax } from 'src/libs/ajax';
-import { Apps } from 'src/libs/ajax/leonardo/Apps';
 import colors from 'src/libs/colors';
 import { FormLabel } from 'src/libs/forms';
-import * as Nav from 'src/libs/nav';
-import { notify } from 'src/libs/notifications';
-import { getUser } from 'src/libs/state';
 import * as Utils from 'src/libs/utils';
 import { withBusyState } from 'src/libs/utils';
-import { resolveRunningCromwellAppUrl } from 'src/libs/workflows-app-utils';
 import { getMethodVersionName } from 'src/workflows-app/utils/method-common';
 import validate from 'validate.js';
 
@@ -27,38 +21,7 @@ const constraints = {
   },
 };
 
-const submitMethod = async (signal, onDismiss, method, workspace) => {
-  const namespace = await workspace.workspace.namespace;
-  try {
-    const cbasUrl = (
-      await Apps(signal)
-        .listAppsV2(workspace.workspace.workspaceId)
-        .then((apps) => resolveRunningCromwellAppUrl(apps, getUser()?.email))
-    ).cbasUrl;
-
-    if (cbasUrl) {
-      const methodPayload = {
-        method_name: method.method_name,
-        method_description: method.method_description,
-        method_source: method.method_source,
-        method_version: method.method_version,
-        method_url: method.method_url,
-      };
-      const methodObject = await Ajax(signal).Cbas.methods.post(cbasUrl, methodPayload);
-      onDismiss();
-      Nav.goToPath('workspace-workflows-app-submission-config', {
-        name: workspace.workspace.name,
-        namespace,
-        methodId: methodObject.method_id,
-      });
-    }
-  } catch (error) {
-    notify('error', 'Error creating new method', { detail: error instanceof Response ? await error.text() : error });
-    onDismiss();
-  }
-};
-
-const ImportGithub = ({ setLoading, signal, onDismiss, workspace }) => {
+const ImportGithub = ({ setLoading, signal, onDismiss, workspace, submitMethod }) => {
   const [methodName, setMethodName] = useState('');
   const [methodVersionName, setMethodVersionName] = useState('');
   const [methodUrl, setMethodUrl] = useState('');
